Shift leave team button position, and fix sorting own user as first in list
