+++ apps/studio/components/interfaces/Organization/TeamSettings/MembersView.tsx
@@ -7,6 +7,8 @@ import { GenericSkeletonLoader } from 'components/ui/ShimmeringLoader'
 import { useOrganizationRolesV2Query } from 'data/organization-members/organization-roles-query'
 import { useOrganizationMembersQuery } from 'data/organizations/organization-members-query'
 import { useProfile } from 'lib/profile'
+import { partition } from 'lodash'
+import { useMemo } from 'react'
 import { Button, Loading, Tooltip, TooltipContent, TooltipTrigger } from 'ui'
 import { Admonition } from 'ui-patterns'
 import { MemberRow } from './MemberRow'
@@ -20,7 +22,7 @@ const MembersView = ({ searchString }: MembersViewProps) => {
   const { profile } = useProfile()
 
   const {
-    data: members,
+    data: members = [],
     error: membersError,
     isLoading: isLoadingMembers,
     isError: isErrorMembers,
@@ -35,11 +37,10 @@ const MembersView = ({ searchString }: MembersViewProps) => {
     slug,
   })
 
-  const allMembers = members ?? []
-  const filteredMembers = (
-    !searchString
-      ? allMembers
-      : allMembers.filter((member) => {
+  const filteredMembers = useMemo(() => {
+    return !searchString
+      ? members
+      : members.filter((member) => {
           if (member.invited_at) {
             return member.primary_email?.includes(searchString)
           }
@@ -49,15 +50,17 @@ const MembersView = ({ searchString }: MembersViewProps) => {
             )
           }
         })
-  )
-    .slice()
-    .sort((a, b) => {
-      // [Joshen] Have own account show up top
-      if (a.primary_email === profile?.primary_email) return -1
-      return a.username.localeCompare(b.username)
-    })
+  }, [members, searchString])
 
-  const userMember = allMembers.find((m) => m.primary_email === profile?.primary_email)
+  const [[user], otherMembers] = partition(
+    filteredMembers,
+    (m) => m.primary_email === profile?.primary_email
+  )
+  const sortedMembers = otherMembers.sort((a, b) =>
+    (a.primary_email ?? '').localeCompare(b.primary_email ?? '')
+  )
+
+  const userMember = members.find((m) => m.primary_email === profile?.primary_email)
   const orgScopedRoleIds = (roles?.org_scoped_roles ?? []).map((r) => r.id)
   const isOrgScopedRole = orgScopedRoleIds.includes(userMember?.role_ids?.[0] ?? -1)
 
@@ -117,7 +120,8 @@ const MembersView = ({ searchString }: MembersViewProps) => {
                       </Table.tr>,
                     ]
                   : []),
-                ...filteredMembers.map((member) => (
+                ...(!!user ? [<MemberRow key={user.gotrue_id} member={user} />] : []),
+                ...sortedMembers.map((member) => (
                   <MemberRow key={member.gotrue_id} member={member} />
                 )),
                 ...(searchString.length > 0 && filteredMembers.length === 0
@@ -138,7 +142,7 @@ const MembersView = ({ searchString }: MembersViewProps) => {
                   <Table.td colSpan={12}>
                     <p className="text-foreground-light">
                       {searchString ? `${filteredMembers.length} of ` : ''}
-                      {allMembers.length || '0'} {allMembers.length == 1 ? 'user' : 'users'}
+                      {members.length || '0'} {members.length == 1 ? 'user' : 'users'}
                     </p>
                   </Table.td>
                 </Table.tr>,
